Translations backport for F23
This commit is not useful on its own but may be useful if you want
to backport the translation updates from master to F23 branch.
Here are some instructions.

Normally all translation patches against master should also cleanly
apply to the F23 branch. If they don't:

- copy the most recent *.po files from master to f23,
- apply (cherry-pick) this commit on top of it,
- squash those two commits,
- you get the backported copy of translation updates.

Alternatively you can:

- go to f23 branch,
- apply (cherry-pick) this commit in reverse, the result should be
  the same as master branch,
- apply the translation updates from master, they should apply cleanly,
- apply (cherry-pick) this commit straight, or revert the reversed
  commit you did before,
- squash those three commits,
- you get the backported copy of translation updates.

Always double check the results. If nothing works you have to backport
manually and consider reworking this commit so it works in future.